+++ packages/runtime/src/runtime/hooks/useEnsembleAction.tsx
@@ -162,14 +162,13 @@ export const useExecuteCode: EnsembleActionHook<
                 invokeAPI: async (
                   apiName: string,
                   apiInputs?: { [key: string]: unknown },
-                ) => {
-                  const apiRes = await invokeAPI(
-                    screenData,
-                    apiName,
-                    apiInputs,
-                  );
-                  return apiRes;
-                },
+                ) =>
+                  invokeAPI(screenData, apiName, apiInputs, {
+                    ...customScope,
+                    ensemble: {
+                      env: appContext?.env,
+                    },
+                  }),
                 navigateBack: (): void => navigateBack(navigate),
                 navigateExternalScreen: (url: NavigateExternalScreen) =>
                   navigateExternalScreen(url),
@@ -224,10 +223,11 @@ export const useInvokeAPI: EnsembleActionHook<InvokeAPIAction> = (action) => {
   const [isLoading, setIsLoading] = useState<boolean>();
   const [context, setContext] = useState<{ [key: string]: unknown }>();
   const evaluatedInputs = useEvaluate(action?.inputs, { context });
+  const evaluatedName = useEvaluate({ name: action?.name }, { context });
 
   const api = useMemo(
-    () => apis?.find((model) => model.name === action?.name),
-    [action?.name, apis],
+    () => apis?.find((model) => model.name === evaluatedName.name),
+    [evaluatedName.name, apis],
   );
 
   const onInvokeAPIResponseAction = useEnsembleAction(action?.onResponse);